tombala game
In this project, a multiplayer game known as "Tombola" was designed using a linked list data structure. Tombola is a popular game
It is a game of chance played with numbered cards. The game involves random selection of numbers and players
If the numbers searched match, these numbers are marked on the cards